// client-cargo-main (1)/client-cargo-main/src/stores/Corporate/Corporate.interface.ts
export const IODepartmentEmployee = t.type({
  id: t.string,
  firstName: t.string,
  lastName: t.string,
  personnelNumber: t.string,
});

export type TDepartmentEmployee = t.TypeOf<typeof IODepartmentEmployee>;

export type IDepartments = Record<string, IDepartment[]>;

export interface IDepartment {
  id?: string;
  humanReadableId?: string;
  departmentHeadId?: string;
  parentId?: string;
  location?: string;
  code: string;
  organizationId: string;
  departmentName: string;
  fullStructurePath: string;
  children: IDepartment[];
  employees: TDepartmentEmployee[];
  status: EmployeeStatus;
}

export const IODepartment: t.Type<IDepartment> = t.recursion('IODepartment', () => t.intersection([
  t.type({
    organizationId: t.string,
    code: t.string,
    departmentName: t.string,
    fullStructurePath: t.string,
    employees: t.array(IODepartmentEmployee),
    children: t.array(IODepartment),
    status: ioTypeFromEnum<EmployeeStatus>('EmployeeStatus', EmployeeStatus),
  }),
  t.partial({
    id: t.string,
    humanReadableId: t.string,
    departmentHeadId: t.string,
    parentId: t.string,
    location: t.string,
  }),
])
);

export type TDepartment = t.TypeOf<typeof IODepartment>;